% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readMarkdownTable-package.R, R/readMarkdownTable.R
\docType{package}
\name{readMarkdownTable}
\alias{read-markdown-table}
\alias{read-md-table}
\alias{readMarkdownTable}
\alias{readMarkdownTable-package}
\alias{readMdTable}
\title{readMarkdownTable - read a markdown formatted table into a data set}
\usage{
readMarkdownTable(file, start, end)
}
\arguments{
\item{file}{Input file}

\item{start}{the line number where the table header starts}

\item{end}{the last line in the table (last observation)}
}
\description{
readMarkdownTable - read a markdown formatted table into a data set

This package will read a markdown formatted table into a data frame.
The user must provide the input file and line range of the table.
}
\note{
the file should be given as a character vector (not a connection)
}
\author{
Stefano daros

Stefano Da Ros
}

